# docs/api-spec.yaml
openapi: 3.0.0
info:
  title: AI-Aided Caches n Logs Monitoring n Wiping Daemon API
  version: 1.0.0
  description: A command-line tool for monitoring and wiping cache and log files.
servers:
  - url: http://localhost:8087
paths:
...
  /monitored-path:
    post:
      summary: Adds a path to the monitoring database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: The absolute path to the directory you want to monitor.
                mode:
                  type: string
                  description: Specifies the monitoring mode (log, cache, full, pattern). Default is full.
                  enum:
                    - log
                    - cache
                    - full
                    - pattern
                  default: full
                pattern:
                  type: string
                  description: Specifies the regex pattern to match files (required if mode is pattern).
                noSubdirs:
                  type: boolean
                  description: Prevents monitoring subdirectories (default is false).
                  default: false
              required:
                - path
      responses:
        '201':
          description: Path successfully added to the monitoring database.
        '400':
          description: Invalid input.
        '422':
          description: Invalid or non-existent path.
        '409':
          description: Path already exists in the monitoring database.
        '500':
          description: Internal server error.
    get:
      summary: Returns all monitored paths.
      responses:
        '200':
          description: Returns all monitored paths.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonitorPath'
  /monitored-path/{id}:
    delete:
      summary: Removes a path from the monitoring database.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the path you want to remove.
          schema:
            type: integer
      responses:
        '204':
          description: Path successfully removed from the monitoring database.
        '404':
          description: No monitored path with the specified ID found.
...
components:
  schemas:
...
    MonitorPath:
      type: object
      properties:
        id:
          type: integer
          example: 3210
        path:
          type: string
          example: "/etc/path/to/monitor"
        createdAt:
          type: string
          format: date
          example: "2024-11-01"
...